objId is size_t and compulsory
git-svn-id: file:///home/roberto/Git/svn@7158 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: a7d4e011d6a3b0db549a53f2500113ece956e1e1

+++ trunk/xmipp/applications/programs/window/window_main.cpp
@@ -193,7 +193,7 @@ public:
         Iin.write(fnImgOut);
     }
 
-    void processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+    void processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
     {
         Image<double> img;
         img.read(fnImg,false);
